+++ src/geometry/collider.rs
@@ -113,9 +113,9 @@ pub struct Sensor;
 #[reflect(Component, PartialEq)]
 pub enum ColliderMassProperties {
     /// The mass-properties are computed automatically from the collider’s shape and this density.
-    Density(f32),
+    Density(f64),
     /// The mass-properties are computed automatically from the collider’s shape and this mass.
-    Mass(f32),
+    Mass(f64),
     /// The mass-properties of the collider are replaced by the ones specified here.
     MassProperties(MassProperties),
 }
@@ -134,7 +134,7 @@ pub struct Friction {
     ///
     /// The greater the value, the stronger the friction forces will be.
     /// Should be `>= 0`.
-    pub coefficient: f32,
+    pub coefficient: f64,
     /// The rule applied to combine the friction coefficients of two colliders in contact.
     pub combine_rule: CoefficientCombineRule,
 }
@@ -151,7 +151,7 @@ impl Default for Friction {
 impl Friction {
     /// Creates a `Friction` component from the given friction coefficient, and using the default
     /// `CoefficientCombineRule::Average` coefficient combine rule.
-    pub const fn new(coefficient: f32) -> Self {
+    pub const fn new(coefficient: f64) -> Self {
         Self {
             coefficient,
             combine_rule: CoefficientCombineRule::Average,
@@ -160,7 +160,7 @@ impl Friction {
 
     /// Creates a `Friction` component from the given friction coefficient, and using the default
     /// `CoefficientCombineRule::Average` coefficient combine rule.
-    pub const fn coefficient(coefficient: f32) -> Self {
+    pub const fn coefficient(coefficient: f64) -> Self {
         Self {
             coefficient,
             combine_rule: CoefficientCombineRule::Average,
@@ -176,7 +176,7 @@ pub struct Restitution {
     ///
     /// The greater the value, the stronger the restitution forces will be.
     /// Should be `>= 0`.
-    pub coefficient: f32,
+    pub coefficient: f64,
     /// The rule applied to combine the friction coefficients of two colliders in contact.
     pub combine_rule: CoefficientCombineRule,
 }
@@ -184,7 +184,7 @@ pub struct Restitution {
 impl Restitution {
     /// Creates a `Restitution` component from the given restitution coefficient, and using the default
     /// `CoefficientCombineRule::Average` coefficient combine rule.
-    pub const fn new(coefficient: f32) -> Self {
+    pub const fn new(coefficient: f64) -> Self {
         Self {
             coefficient,
             combine_rule: CoefficientCombineRule::Average,
@@ -193,7 +193,7 @@ impl Restitution {
 
     /// Creates a `Restitution` component from the given restitution coefficient, and using the default
     /// `CoefficientCombineRule::Average` coefficient combine rule.
-    pub const fn coefficient(coefficient: f32) -> Self {
+    pub const fn coefficient(coefficient: f64) -> Self {
         Self {
             coefficient,
             combine_rule: CoefficientCombineRule::Average,
@@ -461,11 +461,11 @@ impl From<ActiveEvents> for rapier::pipeline::ActiveEvents {
 /// The total force magnitude beyond which a contact force event can be emitted.
 #[derive(Copy, Clone, PartialEq, Component, Reflect, FromReflect)]
 #[reflect(Component)]
-pub struct ContactForceEventThreshold(pub f32);
+pub struct ContactForceEventThreshold(pub f64);
 
 impl Default for ContactForceEventThreshold {
     fn default() -> Self {
-        Self(f32::MAX)
+        Self(f64::MAX)
     }
 }
